code/requirements.txt
autogen-agentchat[llm,gemini]~=0.2
flaml[autogen]
streamlit
pydantic
fastapi
pdfminer.six
pymupdf